// src/app/interfaces/pic-gudang/balancing/i-report-balancing.ts
export interface IReportBalancing {
    idDetailBalancing: number;
    kodeBarang: string;
    namaBarang: string;
    stokAwal: number;
    stokAhkhir: number;
    barangIn: number;
    barangOut: number;
    opr: number;
    uc: number;
    nc: number;
    kkb: number;
    ds: number;
    asr: number;
}